refactor(LLM): revert hook movement
